solvers: Moved a test case from test_ode.py to test_systems.py
Moved the test case which was now being handled by the new solver.
This test case is still slow and thus the correct version of the
test case given the new solver was used.
